Initial source (#28)
* Adding initial code.

Project was created in jetbrains WebStorm, but should be runnable on any
system with npm.

An example timeline is on the Timeline page. This timeline was create
with chart.js and the chart.js wrapper for react. Right now it only
populates the timeline with random data.

* Added Dropdown menu and random colour selection for chart items.

* The timeline can now read a CSV.

Dynamic parsing is not implemented yet.

* Made changes to handle exceptionally large datasets.

This prototype will have to be largely scrapped though.

* Updated package-lock.json

* Adding wikidocs and webstorm guide images

* Updated README.md to fit the project template.