Timeline types (#136)
* Renamed ViewType.occurrence to ViewType.event to match our project
specs.

Added a scale property to the TimelineModel. All uses of the private
scale property in TimelineComponent will be removed.

Added TimelineTypeInterface: an interface that allows us to extract and
implement functions specific to drawing and manipulating each type of
timeline.

* Moving TimelineTypeInterface to TimelineTypes directory. The classes for
the four TimelineTypes will also reside in this directory in the future.

Replaced the usage of the old drawEventMagnitude and drawEventInterval
with TimelineTypeInterface calls.

Replaced the usage of the old getEventMagnitudeData and
getEventIntervalData with TimelineTypeInterface calls.

* Added EventMagnitude and IntervalMagnitude classes.

Replaced the drawLabels method in TimelineComponent with drawLabels
method in TimelineTypeInterface.

* Implemented a TimelineType abstract class.

Benefits:
 - Can inherit member functions that should logically be
   handled by the TimelineType, but that are the same for all of them.

 - Can declare abstract methods that must be implemented by the child
   classes, without defining them in TimelineType

 - No objects from TimelineType can be instantiated, only classes that
   extend it.

* Got rid of a bad inspection in TimelineTypeInterface.

Removed some unnecessary test conditions.